requirements.txt
faster-whisper==1.0.3
praat-parselmouth==0.4.3
librosa==0.10.2.post1
numpy>=1.23
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
pyphen==0.14.0
numba==0.59.1
uvicorn==0.30.1
requests==2.32.3